Added instructions to add algorithms
