+++ nevergrad/benchmark/experiments.py
@@ -4,35 +4,14 @@
 # LICENSE file in the root directory of this source tree.
 
 from typing import Iterator, Optional
-import numpy as np
 from ..functions import ArtificialFunction
-from ..common import decorators
 from .. import optimization
 from .xpbase import Experiment
 from .xpbase import create_seed_generator
+from .xpbase import registry
+# register all frozen experiments
+from . import frozenexperiments  # pylint:disable=unused-import
 # pylint: disable=stop-iteration-return
-
-
-registry = decorators.Registry()
-
-
-@registry.register
-def basic(seed: Optional[int] = None) -> Iterator[Experiment]:
-    """Test settings
-    """
-    seedg = create_seed_generator(seed)
-    function = ArtificialFunction(name="sphere", block_dimension=2, noise_level=1)
-    np.random.seed(seed)  # seed before initializing the function!
-    function.initialize()  # initialization uses randomness
-    return iter([Experiment(function, optimizer_name="OnePlusOne", num_workers=2, budget=4, seed=next(seedg))])
-
-
-@registry.register
-def repeated_basic(seed: Optional[int] = None) -> Iterator[Experiment]:
-    """Test settings
-    """
-    seedg = create_seed_generator(seed)
-    return (next(basic(next(seedg))) for _ in range(10))
 
 
 @registry.register
@@ -54,22 +33,6 @@ def discrete(seed: Optional[int] = None) -> Iterator[Experiment]:
 
 
 @registry.register
-def small_discrete(seed: Optional[int] = None) -> Iterator[Experiment]:
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
-    names = ["hardonemax5", "hardjump5", "hardleadingones5"]
-    optims = sorted(x for x, y in optimization.registry.items() if "iscrete" in x and "epea" not in x and "DE" not in x
-                    and "SSNEA" not in x)
-    functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
-                 for name in names for bd in [30] for uv_factor in [5, 10] for n_blocks in [1]]
-    # functions are not initialized and duplicated at yield time, they will be initialized in the experiment (no need to seed here)
-    for func in functions:
-        for optim in optims:
-            for budget in [100, 400, 700, 1000, 1300, 1600, 1900, 2200, 2500, 2800, 3000]:  # , 10000]:
-                yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
-
-
-@registry.register
 def minidoe(seed: Optional[int] = None) -> Iterator[Experiment]:
     # prepare list of parameters to sweep for independent variables
     seedg = create_seed_generator(seed)
@@ -86,69 +49,11 @@ def minidoe(seed: Optional[int] = None) -> Iterator[Experiment]:
 
 
 @registry.register
-def illcond(seed: Optional[int] = None) -> Iterator[Experiment]:
-    """All optimizers on ill cond problems
-    """
-    seedg = create_seed_generator(seed)
-    for budget in [500, 1000, 2000, 4000]:
-        for optim in ["SQP", "DE", "CMA", "PSO", "RotationInvariantDE", "NelderMead"]:
-            for rotation in [True, False]:
-                for name in ["ellipsoid", "cigar"]:
-                    function = ArtificialFunction(name=name, rotation=rotation, block_dimension=100)
-                    yield Experiment(function, optim, budget=budget, seed=next(seedg))
-
-
-@registry.register
-def compabasedillcond(seed: Optional[int] = None) -> Iterator[Experiment]:
-    """All optimizers on ill cond problems
-    """
-    seedg = create_seed_generator(seed)
-    for budget in [500, 1000, 2000, 4000, 8000]:
-        for optim in ["DE", "CMA", "PSO", "BPRotationInvariantDE", "RotationInvariantDE",
-                      "AlmostRotationInvariantDE", "AlmostRotationInvariantDEAndBigPop"]:
-            for rotation in [True, False]:
-                for name in ["ellipsoid", "cigar"]:
-                    function = ArtificialFunction(name=name, rotation=rotation, block_dimension=30)
-                    yield Experiment(function, optim, budget=budget, seed=next(seedg))
-
-
-@registry.register
-def noise(seed: Optional[int] = None) -> Iterator[Experiment]:
-    """All optimizers on ill cond problems
-    """
-    seedg = create_seed_generator(seed)
-    optims = sorted(x for x, y in optimization.registry.items()
-                    if ("TBPSA" in x or "ois" in x or "CMA" in x or "epea" in x) and "iscr" not in x)
-    for budget in [500, 1000, 2000, 4000, 8000, 16000, 32000, 64000, 128000 ]:
-        for optim in optims:
-            for rotation in [True, False]:
-                for name in ["sphere", "cigar", "sphere4"]:
-                    function = ArtificialFunction(name=name, rotation=rotation, block_dimension=20, noise_level=10)
-                    yield Experiment(function, optim, budget=budget, seed=next(seedg))
-
-
-@registry.register
-def doe_dim4(seed: Optional[int] = None) -> Iterator[Experiment]:  # Here, QR performs best, then Random, then LHS, then Cauchy.
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
-    names = ["sphere"]  # n for n in ArtificialFunction.list_sorted_function_names() if "sphere" in n]
-    optims = sorted(x for x, y in optimization.registry.items() if y.one_shot and "arg" not in x and "mal" not in x)
-    functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
-                 for name in names for bd in [4] for uv_factor in [0] for n_blocks in [1]]
-    # functions are not initialized and duplicated at yield time, they will be initialized in the experiment (no need to seed here)
-    for func in functions:
-        for optim in optims:
-            for budget in [30, 100, 3000, 10000]:
-                # duplicate -> each Experiment has different randomness
-                yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
-
-
-@registry.register
 def doe_dim10(seed: Optional[int] = None) -> Iterator[Experiment]:  # LHS performs best, followed by QR and random
     # nearly equally (Hammersley better than random, Halton not clearly; scrambling improves results).
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
+    # prepare list of parameters to sweep for independant variables
     names = ["sphere"]
+    seedg = create_seed_generator(seed)
     optims = sorted(x for x, y in optimization.registry.items() if y.one_shot and "arg" not in x and "mal" not in x)
     functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
                  for name in names for bd in [10] for uv_factor in [0] for n_blocks in [1]]
@@ -156,53 +61,5 @@ def doe_dim10(seed: Optional[int] = None) -> Iterator[Experiment]:  # LHS perfor
     for func in functions:
         for optim in optims:
             for budget in [30, 100, 3000, 10000]:
-                # duplicate -> each Experiment has different randomness
-                yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
-
-
-@registry.register
-def dim10_smallbudget(seed: Optional[int] = None) -> Iterator[Experiment]:
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
-    names = ["sphere"]
-    optims = sorted(x for x, y in optimization.registry.items() if y.one_shot and "arg" not in x and "mal" not in x)
-    functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
-                 for name in names for bd in [10] for uv_factor in [0] for n_blocks in [1]]
-    # functions are not initialized and duplicated at yield time, they will be initialized in the experiment (no need to seed here)
-    for func in functions:
-        for optim in optims:
-            for budget in [4, 8, 16, 32]:
-                # duplicate -> each Experiment has different randomness
-                yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
-
-
-@registry.register
-def dim10_select_two_features(seed: Optional[int] = None) -> Iterator[Experiment]:     # 2 variables matter - Scrambled Hammersley rules.
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
-    names = ["sphere"]
-    optims = sorted(x for x, y in optimization.registry.items() if y.one_shot and "arg" not in x and "mal" not in x)
-    functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
-                 for name in names for bd in [2] for uv_factor in [5] for n_blocks in [1]]
-    # functions are not initialized and duplicated at yield time, they will be initialized in the experiment (no need to seed here)
-    for func in functions:
-        for optim in optims:
-            for budget in [4, 8, 16, 32]:
-                # duplicate -> each Experiment has different randomness
-                yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
-
-
-@registry.register
-def dim10_select_one_feature(seed: Optional[int] = None) -> Iterator[Experiment]:    # One and only one variable matters - LHS wins.
-    # prepare list of parameters to sweep for independent variables
-    seedg = create_seed_generator(seed)
-    names = ["sphere"]
-    optims = sorted(x for x, y in optimization.registry.items() if y.one_shot and "arg" not in x and "mal" not in x)
-    functions = [ArtificialFunction(name, block_dimension=bd, num_blocks=n_blocks, useless_variables=bd * uv_factor * n_blocks)
-                 for name in names for bd in [1] for uv_factor in [10] for n_blocks in [1]]
-    # functions are not initialized and duplicated at yield time, they will be initialized in the experiment (no need to seed here)
-    for func in functions:
-        for optim in optims:
-            for budget in [8, 10, 12, 14, 16, 18, 20]:
                 # duplicate -> each Experiment has different randomness
                 yield Experiment(func.duplicate(), optim, budget=budget, num_workers=1, seed=next(seedg))
